Move JsonComponent docs to general Jackson section
Closes gh-30405
