+++ spring-boot-project/spring-boot-docs/src/docs/asciidoc/anchor-rewrite.properties
@@ -1013,3 +1013,6 @@ dependency-versions-properties=appendix.dependency-versions.properties
 
 dependency-versions.coordinates=appendix.dependency-versions.coordinates
 dependency-versions.properties=appendix.dependency-versions.properties
+
+# gh-30405
+web.servlet.spring-mvc.json=features.json.jackson.custom-serializers-and-deserializers
