+++ spring-boot-project/spring-boot-docs/src/docs/asciidoc/web/servlet.adoc
@@ -75,28 +75,6 @@ You can also override default converters in the same way.
 
 
 
-[[web.servlet.spring-mvc.json]]
-==== Custom JSON Serializers and Deserializers
-If you use Jackson to serialize and deserialize JSON data, you might want to write your own `JsonSerializer` and `JsonDeserializer` classes.
-Custom serializers are usually https://github.com/FasterXML/jackson-docs/wiki/JacksonHowToCustomSerializers[registered with Jackson through a module], but Spring Boot provides an alternative `@JsonComponent` annotation that makes it easier to directly register Spring Beans.
-
-You can use the `@JsonComponent` annotation directly on `JsonSerializer`, `JsonDeserializer` or `KeyDeserializer` implementations.
-You can also use it on classes that contain serializers/deserializers as inner classes, as shown in the following example:
-
-include::code:MyJsonComponent[]
-
-All `@JsonComponent` beans in the `ApplicationContext` are automatically registered with Jackson.
-Because `@JsonComponent` is meta-annotated with `@Component`, the usual component-scanning rules apply.
-
-Spring Boot also provides {spring-boot-module-code}/jackson/JsonObjectSerializer.java[`JsonObjectSerializer`] and {spring-boot-module-code}/jackson/JsonObjectDeserializer.java[`JsonObjectDeserializer`] base classes that provide useful alternatives to the standard Jackson versions when serializing objects.
-See {spring-boot-module-api}/jackson/JsonObjectSerializer.html[`JsonObjectSerializer`] and {spring-boot-module-api}/jackson/JsonObjectDeserializer.html[`JsonObjectDeserializer`] in the Javadoc for details.
-
-The example above can be rewritten to use `JsonObjectSerializer`/`JsonObjectDeserializer` as follows:
-
-include::code:object/MyJsonComponent[]
-
-
-
 [[web.servlet.spring-mvc.message-codes]]
 ==== MessageCodesResolver
 Spring MVC has a strategy for generating error codes for rendering error messages from binding errors: `MessageCodesResolver`.
