+++ spring-boot-project/spring-boot-docs/src/main/java/org/springframework/boot/docs/features/json/jackson/customserializersanddeserializers/MyObject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2022 the original author or authors.
+ * Copyright 2012-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,20 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.docs.web.servlet.springmvc.json.`object`
+package org.springframework.boot.docs.features.json.jackson.customserializersanddeserializers;
 
-class MyObject(val name: String = "", val age: Int = 0)
+class MyObject {
+
+	MyObject(String name, int age) {
+	}
+
+	String getName() {
+		return null;
+	}
+
+	Integer getAge() {
+		return null;
+
+	}
+
+}
