+++ spring-boot-project/spring-boot-docs/src/main/java/org/springframework/boot/docs/features/json/jackson/customserializersanddeserializers/object/MyObject.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.docs.web.servlet.springmvc.json.object;
+package org.springframework.boot.docs.features.json.jackson.customserializersanddeserializers.object;
 
 class MyObject {
